+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Issue #416](https://github.com/influxdb/influxdb/issues/416). Improve the time it takes to drop database
 - [Issue #413](https://github.com/influxdb/influxdb/issues/413). Don't assume that group by interval is greater than a second
 - [Issue #415](https://github.com/influxdb/influxdb/issues/415). Include the database when sending an auth error back to the user
+- [Issue #421](https://github.com/influxdb/influxdb/issues/421). Make read timeout a config option
 
 ### Bugfixes